chore(ops): merge all ops into a single ops.rs
All subcommands are just barely 1k lines of “code”, and we will never
be able to dedup anything if we split them into a million little
files.

Thus, for the sanity of all, let’s merge them. May a thousand
refactors reduce this file in the future.

Fixes https://github.com/nix-community/lorri/issues/15
