+++ tests/daemon/main.rs
@@ -45,7 +45,7 @@ pub fn start_job_with_ping() -> std::io::Result<()> {
 
     connect(&address, Duration::from_millis(1000));
 
-    lorri::ops::ping::main(
+    lorri::ops::ping(
         lorri::NixFile::from(lorri::AbsPathBuf::new(shell_nix).unwrap()),
         Some(address),
     )
